agent: refactor: rename siglip2 to lfm2_siglip2
