+++ vllm/model_executor/models/lfm2_siglip2.py
